requirements.txt
PyJWT~=2.8.0
fastapi~=0.89.1
passlib~=1.7.4
jose~=1.0.0
SQLAlchemy~=2.0.27
uvicorn~=0.20.0
pydantic~=1.10.14